Includes physics manager and all code working
Have skybox. Need to make animation and make the cannonballs and place
them and the hat and add the code to process the multiple collision
shapes for one object and add them all and then figure out how to make
lua scripts able to call physics functions and apply to dodge buttons
and make scroll bar to adjust force. Figure out how to make boat float.